upload images and set pdf

+++ controllers/adminControllers.js
@@ -1,6 +1,6 @@
 const Product = require("../models/product");
 const { validationResult } = require("express-validator");
-
+const fileHelper = require("../util/file");
 exports.getAddProduct = (req, res, next) => {
     res.render("admin/edit-product", {
         pageTitle: "Add Product",
@@ -14,11 +14,26 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
     const errors = validationResult(req);
-    console.log(errors);
+    // console.log(image);
+    if (!image) {
+        return res.status(422).render("admin/edit-product", {
+            pageTitle: "Add Product",
+            path: "/admin/adit-product",
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description,
+            },
+            errorMessage: "Attached file is not an image",
+            validationErrors: [],
+        });
+    }
     if (!errors.isEmpty()) {
         return res.status(422).render("admin/edit-product", {
             pageTitle: "Add Product",
@@ -35,6 +50,7 @@ exports.postAddProduct = (req, res, next) => {
             validationErrors: errors.array(),
         });
     }
+    const imageUrl = image.path;
     const product = new Product({
         title: title,
         price: price,
@@ -105,7 +121,7 @@ exports.postEditProduct = (req, res, next) => {
     const id = req.body.productId;
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedDesc = req.body.description;
     // Product.findOneAndUpdate(
     //     { _id: id },
@@ -117,8 +133,11 @@ exports.postEditProduct = (req, res, next) => {
     //     },
     //     { new: true }
     // )
+    console.log(image);
+    const errors = validationResult(req);
 
     if (!errors.isEmpty()) {
+        console.log("In error empty");
         return res.status(422).render("admin/edit-product", {
             pageTitle: "Edit Product",
             path: "/admin/edit-product",
@@ -126,7 +145,6 @@ exports.postEditProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDesc,
                 _id: prodId,
@@ -137,13 +155,20 @@ exports.postEditProduct = (req, res, next) => {
     }
     Product.findById(id)
         .then((product) => {
+            console.log(product);
             if (product.userId.toString() !== req.user._id.toString()) {
                 return res.redirect("/");
             }
             product.title = updatedTitle;
             product.price = updatedPrice;
             product.description = updatedDesc;
-            product.price = updatedPrice;
+            if (image) {
+                console.log("Image existed");
+                fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = image.path;
+            }
+            console.log(product);
+
             return product.save().then((result) => {
                 console.log("Updated product");
                 res.redirect("/admin/products");
@@ -177,7 +202,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+    Product.findById(proId)
+        .then((product) => {
+            if (!product) {
+                return next(new Error("product not found"));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: prodId, userId: req.user._id });
+        })
 
         .then(() => {
             console.log("Destroyed Product");
